+++ osu.Game.Tournament/Screens/Editors/TeamEditorScreen.cs
@@ -151,7 +151,7 @@ namespace osu.Game.Tournament.Screens.Editors
                             {
                                 Width = 0.2f,
                                 Margin = new MarginPadding(10),
-                                Text = "编辑随机结果",
+                                Text = "编辑选手介绍信息",
                                 Action = () =>
                                 {
                                     sceneManager?.SetScreen(new SeedingEditorScreen(team));