+++ osu.Game.Tournament/Screens/MapPool/MapPoolScreen.cs
@@ -235,7 +235,7 @@ namespace osu.Game.Tournament.Screens.MapPool
                     {
                         Anchor = Anchor.TopCentre,
                         Origin = Anchor.TopCentre,
-                        Height = 42,
+                        Height = 44,
                     });
                 }
             }